refactor(bridge): split code into modules
